♻️ 画像生成のコンポーネント整理
